+++ R/bb_labs.R
@@ -1,10 +1,11 @@
 #' Quickly set title, axis labels and caption
 #'
 #' @param bb A \code{billboard} \code{htmlwidget} object.
-#' @param title Plot title.
-#' @param x Label for x axis.
-#' @param y Label for y axis.
-#' @param caption Caption for the chart.
+#' @param title Text for the chart title, use \code{\\n} to make a new line.
+#' @param x Text for x axis title.
+#' @param y Text for y axis title.
+#' @param caption Text for the caption displayed in the bottom-right of the chart.
+#' @param caption_href Associate the caption with a link to an URL.
 #'
 #' @return A \code{billboard} \code{htmlwidget} object.
 #' @export
@@ -15,13 +16,18 @@
 #' data("prod_par_filiere")
 #' 
 #' billboarder() %>%
-#'   bb_barchart(data = prod_par_filiere[, c("annee", "prod_hydraulique")], color = "#102246") %>%
-#'   bb_legend(show = FALSE) %>% 
-#'   bb_labs(title = "French hydraulic production", 
-#'           y = "production (in terawatt-hours)",
-#'           caption = "Data source: RTE (https://opendata.rte-france.com)")
-#' 
-bb_labs <- function(bb, title = NULL, x = NULL, y = NULL, caption = NULL) {
+#'   bb_barchart(
+#'     data = prod_par_filiere[, c("annee", "prod_hydraulique")],
+#'     color = "#102246"
+#'   ) %>%
+#'   bb_legend(show = FALSE) %>%
+#'   bb_labs(
+#'     title = "French hydraulic production",
+#'     y = "production (in terawatt-hours)",
+#'     caption = "Data source: RTE (https://opendata.reseaux-energies.fr/)",
+#'     caption_href = "https://opendata.reseaux-energies.fr/"
+#'   )
+bb_labs <- function(bb, title = NULL, x = NULL, y = NULL, caption = NULL, caption_href = NULL) {
   
   if (!is.null(title)) {
     bb <- bb_title(bb, text = title, position = "left-top", 
@@ -38,7 +44,10 @@ bb_labs <- function(bb, title = NULL, x = NULL, y = NULL, caption = NULL) {
   
   if (!is.null(caption)) {
     bb <- .bb_opt(bb, "caption", text = caption)
-    bb <- .bb_opt(bb, "padding", bottom = 20)
+    bb <- .bb_opt(bb, "padding", bottom = 10)
+    if (!is.null(caption)) {
+      bb <- .bb_opt(bb, "caption", href = caption_href)
+    }
   }
   
   return(bb)